api: move getCollData() to utils cdxloader: undo changes to request loading, not needed
